Fixed search colors
