+++ src/components/ContentDecoder.stories.tsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import { ContentDecoder } from './ContentDecoder';
+
+export const Default = () => <div style={{ width: '370px' }}>TBD</div>;
