+++ src/components/PagesMessages.styles.tsx
@@ -1,0 +1,33 @@
+import styled from 'styled-components';
+
+export const LoadingContainer = styled.div`
+    flex: 1;
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    justify-content: center;
+    text-align: center;
+`;
+
+export const Error = styled.div`
+    flex: 1;
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    justify-content: center;
+    text-align: center;
+
+    font-size: 1rem;
+    color: #e96f6f;
+
+    font-size: 1rem;
+    padding: 1rem 1rem;
+    background: #ffffff30;
+    margin-top: 1rem;
+    border-radius: 5.2px;
+`;
+
+export const SadFace = styled.div`
+    font-size: 2rem;
+    margin-bottom: 1rem;
+`;
