+++ src/components/SearchBar.styles.tsx
@@ -1,15 +1,8 @@
 import styled from 'styled-components';
-import * as theme from '../shared/theme';
-import { Severity } from '../types/api';
 
-interface RiskProps {
-    severity: undefined | Severity;
-}
-
-export const Form = styled.form<RiskProps>`
+export const Form = styled.form`
     display: flex;
     flex-direction: row;
-    /* align-items: center; */
     gap: 8px;
     flex: 1;
